Move util functions to separate module
