--- src/features/books/util.ts
@@ -1,5 +1,0 @@
-// Calculate discount rate (%) given original and discounted prices.
-export function calculateBookDiscountRate(discountedPurchaseAmount: number, purchaseAmount: number): number {
-  if (purchaseAmount <= 0) return 0;
-  return Math.round((purchaseAmount - discountedPurchaseAmount) / purchaseAmount * 100);
-}
